+++ benchmarks/SkyriseMini/Consts.cs
