--- benchmarks/SkyriseMini/ProtoActorSut.Server/PingPongActor.cs
@@ -1,12 +1,0 @@
-﻿using Proto;
-using ProtoActorSut.Contracts;
-
-namespace ProtoActorSut.Server;
-
-public class PingPongActor : PingPongActorBase
-{
-    public PingPongActor(IContext context) : base(context) { }
-
-    public override Task<PongMessage> Ping(PingMessage request) =>
-        Task.FromResult(new PongMessage {Response = "Hello " + request.Name});
-}
