--- benchmarks/SkyriseMini/ProtoActorSut.Server/Properties/launchSettings.json
@@ -1,16 +1,0 @@
-﻿{
-  "profiles": {
-    "ProtoActorSut.Server": {
-      "commandName": "Project",
-      "dotnetRunMessages": true,
-      "launchBrowser": false,
-      "applicationUrl": "http://localhost:5900",
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development",
-        "DOTNET_TieredPGO":"1",
-        "DOTNET_TC_QuickJitForLoops":"1",
-        "DOTNET_ReadyToRun":"0"
-      }
-    }
-  }
-}
