+++ benchmarks/SkyriseMini/ProtoActorSut.Shared.csproj
