+++ benchmarks/SkyriseMini/SkyriseMini.csproj
@@ -1,0 +1,39 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net6.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\..\src\Proto.Cluster.Consul\Proto.Cluster.Consul.csproj" />
+      <ProjectReference Include="..\..\src\Proto.Cluster\Proto.Cluster.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <PackageReference Include="Grpc.Tools" Version="2.47.0">
+        <PrivateAssets>all</PrivateAssets>
+        <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+      </PackageReference>
+      <PackageReference Include="OpenTelemetry" Version="1.3.0" />
+      <PackageReference Include="OpenTelemetry.Api" Version="1.3.0" />
+      <PackageReference Include="OpenTelemetry.Exporter.Prometheus" Version="1.3.0-rc.2" />
+      <PackageReference Include="prometheus-net" Version="6.0.0" />
+      <PackageReference Include="Serilog.AspNetCore" Version="6.0.0-dev-00265" />
+      <PackageReference Include="Serilog.Sinks.Console" Version="4.0.2-dev-00890" />
+      <PackageReference Include="Serilog.Sinks.Seq" Version="5.1.2-dev-00225" />
+      <PackageReference Include="Swashbuckle.AspNetCore" Version="6.3.1" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <Protobuf Include="protos.proto" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <Folder Include="Bus" />
+      <Folder Include="ProtoActor" />
+    </ItemGroup>
+
+</Project>
