--- benchmarks/SkyriseMini/TestRunner.Client/Properties/launchSettings.json
@@ -1,18 +1,0 @@
-﻿{
-  "$schema": "https://json.schemastore.org/launchsettings.json",
-  "profiles": {
-    "TestRunner.Client": {
-      "commandName": "Project",
-      "dotnetRunMessages": true,
-      "launchBrowser": true,
-      "launchUrl": "swagger",
-      "applicationUrl": "http://localhost:5051",
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development",
-        "DOTNET_TieredPGO":"1",
-        "DOTNET_TC_QuickJitForLoops":"1",
-        "DOTNET_ReadyToRun":"0"
-      }
-    }
-  }
-}
