--- benchmarks/SkyriseMini/TestRunner.Client/TestRunner.Client.csproj
@@ -1,25 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-    <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
-        <Nullable>enable</Nullable>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <PublishReadyToRun>false</PublishReadyToRun>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="MassTransit.Azure.ServiceBus.Core" Version="8.0.3" />
-        <PackageReference Include="MassTransit.RabbitMQ" Version="8.0.3" />
-        <PackageReference Include="Serilog" Version="2.10.0" />
-        <PackageReference Include="Serilog.AspNetCore" Version="5.0.0" />
-        <PackageReference Include="Serilog.Enrichers.Environment" Version="2.2.0" />
-        <PackageReference Include="Serilog.Settings.Configuration" Version="3.3.0" />
-        <PackageReference Include="Serilog.Sinks.Seq" Version="5.1.1" />
-        <PackageReference Include="Swashbuckle.AspNetCore" Version="6.3.0" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\TestRunner.Contract\TestRunner.Contract.csproj" />
-    </ItemGroup>
-
-</Project>
