--- benchmarks/SkyriseMini/TestRunner.Client/appsettings.Development.json
@@ -1,2 +1,0 @@
-{
-}
