--- benchmarks/SkyriseMini/TestRunner.Client/appsettings.json
@@ -1,25 +1,0 @@
-{
-  "Bus": {
-    "RabbitMq": {
-      "Host": "localhost",
-      "User": "guest",
-      "Password": "guest"
-    },
-
-    "ServiceBusConnectionString": "_service_bus_connection_string_"
-  },
-
-  "SeqUrl": "http://localhost:5341",
-
-  "Serilog": {
-    "MinimumLevel": {
-      "Default": "Information",
-      "Override": {
-        "Microsoft.AspNetCore": "Warning",
-        "Microsoft.Hosting.Lifetime": "Information"
-      }
-    }
-  },
-  
-  "AllowedHosts": "*"
-}
