--- benchmarks/SkyriseMini/TestRunner.Contract/TestRunner.Contract.csproj
@@ -1,10 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <PublishReadyToRun>false</PublishReadyToRun>
-    </PropertyGroup>
-
-</Project>
