--- benchmarks/SkyriseMini/TestRunner/appsettings.json
@@ -1,48 +1,0 @@
-{
-  "ActorFramework": "Orleans",
-
-  "Orleans": {
-    "SiloPort": 11112,
-    "gatewayPort": 30001,
-    "ClusterId": "dev",
-    "ServiceId": "ActorBenchmark",
-    "ClusteringStorage": "_storage_connection_string_"
-  },
-  
-  "ProtoActor": {
-    "ClusterName": "dev"
-  },
-
-  "Akka": {
-    "ActorSystemName": "actor-benchmark",
-    "Role": "runner",
-    "SutRole": "sut",
-    "PublicHostname": "localhost",
-    "Port": 5056,
-    "SeedNodes": ["localhost:5800"]
-  },
-
-  "Bus": {
-    "RabbitMq": {
-      "Host": "localhost",
-      "User": "guest",
-      "Password": "guest"
-    },
-
-    "ServiceBusConnectionString": "_service_bus_connection_string_"
-  },
-  
-  "SeqUrl": "http://localhost:5341",
-
-  "Serilog": {
-    "MinimumLevel": {
-      "Default": "Information",
-      "Override": {
-        "Microsoft.AspNetCore": "Warning",
-        "Microsoft.Hosting.Lifetime": "Information"
-      }
-    }
-  },
-  
-  "AllowedHosts": "*"
-}
