+++ benchmarks/SkyriseMini/Tests/ActivationTest.cs
@@ -1,5 +1,6 @@
 ﻿using System.Diagnostics;
 using System.Threading.Channels;
+using Microsoft.Extensions.Logging;
 using TestRunner.Monitoring;
 
 namespace TestRunner.Tests;
