+++ benchmarks/SkyriseMini/protos.proto
@@ -9,7 +9,3 @@ message PingMessage {
 message PongMessage {
   string response = 1;
 }
-
-service PingPongActor {
-  rpc Ping(PingMessage) returns (PongMessage);
-}
